submit and activate  added

+++ client/app/(main)/components/student_results/page.tsx
@@ -45,6 +45,9 @@ const StudentResultComponent = (props: StudentResultProps) => {
     const deleteStudentResult = async () => {
         try {
             if (selectedStudentResult) {
+                if(selectedStudentResult.status==="CLOSED"){
+                    throw new Error('Closed Results Can not be deleted!');
+                }
                 const deleted = await StudentResultService.deleteStudentResult(selectedStudentResult);
                 if (deleted) {
                     let _studentResults = (studentResults as any)?.filter((val: any) => val._id !== selectedStudentResult._id);
@@ -116,6 +119,7 @@ const StudentResultComponent = (props: StudentResultProps) => {
                         <Column field="subject_weight.assessment_type" header="Assesment" sortable headerStyle={{ minWidth: '10rem' }} />
                         <Column field="subject_weight.assessment_weight" header="Weight" sortable headerStyle={{ minWidth: '10rem' }} />
                         <Column field="result" header="Result" sortable headerStyle={{ minWidth: '10rem' }} />
+                        <Column field="status" header="Status" sortable headerStyle={{ minWidth: '10rem' }}></Column>
                         <Column body={actionBodyTemplate} headerStyle={{ minWidth: '10rem' }} />
                     </DataTable>
